Split ElevatorTest into multiple focussed module tests
