+++ application/src/test/java/org/opentripplanner/graph_builder/module/osm/moduletests/elevator/MultipleLevelsTest.java
@@ -4,17 +4,12 @@ import static com.google.common.truth.Truth.assertThat;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.opentripplanner.osm.model.NodeBuilder.node;
 
-import java.time.Duration;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.CsvSource;
 import org.opentripplanner.core.model.accessibility.Accessibility;
-import org.opentripplanner.graph_builder.issue.service.DefaultDataImportIssueStore;
-import org.opentripplanner.graph_builder.issues.FewerThanTwoIntersectionNodesInElevatorWay;
 import org.opentripplanner.graph_builder.module.osm.OsmModuleTestFactory;
 import org.opentripplanner.osm.TestOsmProvider;
 import org.opentripplanner.osm.model.NodeBuilder;
@@ -31,47 +26,7 @@ import org.opentripplanner.street.model.vertex.OsmEntityType;
 import org.opentripplanner.street.model.vertex.OsmVertex;
 import org.opentripplanner.streetadapter.VertexFactory;
 
-class ElevatorTest {
-
-  @Test
-  void testDuration() {
-    var way = OsmWay.of().withTag("duration", "00:01:02").withTag("highway", "elevator").build();
-    var provider = TestOsmProvider.of().addWay(way).build();
-    var graph = new Graph();
-    var osmModule = OsmModuleTestFactory.of(provider).withGraph(graph).builder().build();
-
-    osmModule.buildGraph();
-
-    var edges = graph.getEdgesOfType(ElevatorHopEdge.class);
-    assertThat(edges).hasSize(2);
-    for (var edge : edges) {
-      assertThat(edge.getTravelTime()).hasValue(Duration.ofSeconds(62));
-    }
-  }
-
-  @Test
-  void testMultilevelNodeDuration() {
-    var node0 = node(0, new WgsCoordinate(0, 0));
-    var node1 = node(1, new WgsCoordinate(2, 0));
-    var elevatorNode = NodeBuilder.of(2, new WgsCoordinate(1, 0))
-      .withTag("duration", "00:01:02")
-      .withTag("highway", "elevator")
-      .withTag("level", "1;2")
-      .build();
-    var provider = TestOsmProvider.of()
-      .addWayFromNodes(way -> way.withTag("level", "1"), node0, elevatorNode)
-      .addWayFromNodes(way -> way.withTag("level", "2"), node1, elevatorNode)
-      .build();
-    var graph = new Graph();
-
-    OsmModuleTestFactory.of(provider).withGraph(graph).builder().build().buildGraph();
-
-    var edges = graph.getEdgesOfType(ElevatorHopEdge.class);
-    assertThat(edges).hasSize(2);
-    for (var edge : edges) {
-      assertThat(edge.getTravelTime()).hasValue(Duration.ofSeconds(62));
-    }
-  }
+class MultipleLevelsTest {
 
   @Test
   void testMultilevelNodeWithWaysOnSameLevel() {
@@ -312,117 +267,6 @@ class ElevatorTest {
       )
     );
     assertEquals(edgeSet.size(), graph.getEdges().size());
-  }
-
-  @ParameterizedTest
-  @CsvSource(
-    value = {
-      "1, 2, 3, 4", "1, 1, 1, 1", "0, 1, 1, null", "null, null, 1, null", "null, null, null, null",
-    },
-    nullValues = "null"
-  )
-  void testOsmElevatorNodeUniqueLabels(String level1, String ref1, String level2, String ref2) {
-    var n1 = node(1, new WgsCoordinate(0, 1));
-    var n2 = node(2, new WgsCoordinate(0, 2));
-    var elevatorNode = NodeBuilder.of(3, new WgsCoordinate(0, 3))
-      .withTag("highway", "elevator")
-      .build();
-    var provider = TestOsmProvider.of()
-      .addWayFromNodes(
-        way -> {
-          way.withTag("level", level1);
-          way.withTag("level:ref", ref1);
-        },
-        n1,
-        elevatorNode
-      )
-      .addWayFromNodes(
-        way -> {
-          way.withTag("level", level2);
-          way.withTag("level:ref", ref2);
-        },
-        elevatorNode,
-        n2
-      )
-      .build();
-    var graph = new Graph();
-
-    OsmModuleTestFactory.of(provider).withGraph(graph).builder().build().buildGraph();
-
-    assertEquals(
-      graph.getVertices().size(),
-      graph
-        .getVertices()
-        .stream()
-        .map(vertex -> vertex.getLabel())
-        .distinct()
-        .count()
-    );
-  }
-
-  /**
-   * If an elevator way has three intersection nodes, it is is probably a tagging error.
-   * OTP supports it anyway. We need to make sure that the middle intersection node is not created
-   * twice, otherwise we get an error during deserialization.
-   */
-  @Test
-  void elevatorWayWithThreeIntersectionNodes() {
-    var n1 = node(1, new WgsCoordinate(1, 1));
-    var n2 = node(2, new WgsCoordinate(2, 2));
-    var n3 = node(3, new WgsCoordinate(3, 3));
-    var n4 = node(4, new WgsCoordinate(4, 4));
-    var n5 = node(5, new WgsCoordinate(5, 5));
-
-    var provider = TestOsmProvider.of()
-      .addWayFromNodes(way -> way.withTag("highway", "elevator"), n1, n2, n3)
-      .addWayFromNodes(way -> way.withTag("public_transport", "platform"), n4, n2, n5)
-      .build();
-    var graph = new Graph();
-
-    OsmModuleTestFactory.of(provider).withGraph(graph).builder().build().buildGraph();
-
-    var elevatorHopEdges = graph.getEdgesOfType(ElevatorHopEdge.class);
-    assertThat(elevatorHopEdges).hasSize(4);
-    var elevatorHopVertices = graph
-      .getVerticesOfType(ElevatorHopVertex.class)
-      .stream()
-      .map(vertex -> vertex.getLabelString());
-    assertThat(elevatorHopVertices).containsNoDuplicates();
-  }
-
-  /**
-   * If the connected nodes of an elevator way have been modeled as elevators, they do not appear
-   * as intersection nodes. OTP should create an issue, but it should not cause errors.
-   */
-  @Test
-  void elevatorWayWithFewerThanTwoIntersectionNodes() {
-    var n1 = NodeBuilder.of(1, new WgsCoordinate(1, 1)).withTag("highway", "elevator").build();
-    var n2 = NodeBuilder.of(2, new WgsCoordinate(2, 2)).withTag("highway", "elevator").build();
-
-    var provider = TestOsmProvider.of()
-      .addWayFromNodes(way -> way.withTag("highway", "elevator"), n1, n2)
-      .build();
-    var graph = new Graph();
-    var issueStore = new DefaultDataImportIssueStore();
-
-    OsmModuleTestFactory.of(provider)
-      .withGraph(graph)
-      .builder()
-      .withIssueStore(issueStore)
-      .build()
-      .buildGraph();
-
-    var elevatorHopEdges = graph.getEdgesOfType(ElevatorHopEdge.class);
-    assertThat(elevatorHopEdges).hasSize(0);
-
-    var issues = issueStore
-      .listIssues()
-      .stream()
-      .filter(issue -> issue instanceof FewerThanTwoIntersectionNodesInElevatorWay)
-      .map(FewerThanTwoIntersectionNodesInElevatorWay.class::cast)
-      .toList();
-    assertEquals(1, issues.size());
-    assertEquals(0, issues.getFirst().intersectionNodes());
   }
 
   private void addElevatorBoardAndAlightEdges(
